jigged some things around to propose into @mikhailswifts PR
